student_template/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
gradio>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
